backend/pyproject.toml
[project]
name = "proxymind-backend"
version = "0.1.0"
description = "ProxyMind backend bootstrap"
requires-python = ">=3.14.3"
dependencies = [
  "alembic>=1.18.4",
  "arq>=0.27.0",
  "asyncpg>=0.31.0",
  "fastapi>=0.135.2",
  "google-cloud-documentai>=3.0.0",
  "google-genai>=1.55.0",
  "httpx>=0.28.1",
  "litellm",
  "opentelemetry-api>=1.38.0",
  "opentelemetry-exporter-otlp-proto-grpc>=1.38.0",
  "opentelemetry-instrumentation-fastapi>=0.59b0",
  "opentelemetry-instrumentation-httpx>=0.59b0",
  "opentelemetry-instrumentation-redis>=0.59b0",
  "opentelemetry-instrumentation-sqlalchemy>=0.59b0",
  "opentelemetry-sdk>=1.38.0",
  "pydantic-settings>=2.12.5",
  "prometheus-client>=0.23.1",
  "pypdf>=6.9.2",
  "python-multipart>=0.0.20",
  "qdrant-client>=1.17.1",
  "redis[hiredis]>=5.3.0,<6",
  "sqlalchemy[asyncio]>=2.0.48",
  "structlog>=25.5.0",
  "tenacity>=9.1.4",
  "tinytag>=2.2.1",
  "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
  "httpx-sse>=0.4.0",
  "pytest>=8.4.2",
  "pytest-asyncio>=1.2.0",
  "pytest-timeout>=2.4.0",
  "ruff>=0.15.7",
  "testcontainers[postgres]>=4.14.2",
]

[tool.ruff]
line-length = 100
target-version = "py314"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.ruff.lint.per-file-ignores]
"migrations/versions/*.py" = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.uv.sources]
litellm = { url = "https://github.com/BerriAI/litellm/archive/refs/tags/v1.82.3-stable.patch.2.zip" }

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.hatch.build.targets.sdist]
include = ["app"]

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"
